multi: move messenger utilities into onion_message module
