+++ src/onion_messenger.rs
@@ -1,0 +1,44 @@
+use lightning::chain::keysinterface::EntropySource;
+use lightning::util::logger::{Level, Logger, Record};
+use log::{debug, error, info, trace, warn};
+use rand_chacha::ChaCha20Rng;
+use rand_core::{RngCore, SeedableRng};
+use std::cell::RefCell;
+
+// MessengerUtilities implements some utilites required for onion messenging.
+pub(crate) struct MessengerUtilities {
+    entropy_source: RefCell<ChaCha20Rng>,
+}
+
+impl MessengerUtilities {
+    pub(crate) fn new() -> Self {
+        MessengerUtilities {
+            entropy_source: RefCell::new(ChaCha20Rng::from_entropy()),
+        }
+    }
+}
+
+impl EntropySource for MessengerUtilities {
+    // TODO: surface LDK's EntropySource and use instead.
+    fn get_secure_random_bytes(&self) -> [u8; 32] {
+        let mut chacha_bytes: [u8; 32] = [0; 32];
+        self.entropy_source
+            .borrow_mut()
+            .fill_bytes(&mut chacha_bytes);
+        chacha_bytes
+    }
+}
+
+impl Logger for MessengerUtilities {
+    fn log(&self, record: &Record) {
+        let args_str = record.args.to_string();
+        match record.level {
+            Level::Gossip => {}
+            Level::Trace => trace!("{}", args_str),
+            Level::Debug => debug!("{}", args_str),
+            Level::Info => info!("{}", args_str),
+            Level::Warn => warn!("{}", args_str),
+            Level::Error => error!("{}", args_str),
+        }
+    }
+}
